fix(professionals): client-only page via dynamic import (ssr:false) to stop prerender errors
